+++ Classes/Adventure.js
@@ -3,6 +3,7 @@ const { MAX_MESSAGE_ACTION_ROWS } = require("../constants.js");
 const Combatant = require("./Combatant.js");
 const Resource = require("./Resource.js");
 const { Room } = require("./Room.js");
+const Delver = require("./Delver.js");
 
 module.exports.Adventure = class {
 	/** This read-write payload class describes an ongoing adventure
@@ -31,6 +32,7 @@ module.exports.Adventure = class {
 		battleRound: ""
 	};
 	leaderId = "";
+	/** @type {Delver[]} */
 	delvers = [];
 	challenges = {}; // {challengeName: {intensity, reward, duration}} 0 = done, null = permanent
 	scouting = {