+++ Source/equipment/battleaxe-thirsting.js
@@ -16,9 +16,9 @@ function effect([target], user, isCrit, adventure) {
 		return ` ${getFullName(target, adventure.room.enemyTitles)} was already dead!`;
 	}
 
-	let { element, modifiers: [elementStagger], damage, bonusDamage, critBonus, healing } = module.exports;
+	let { element, modifiers: [elementStagger], damage, bonus, critBonus, healing } = module.exports;
 	if (user.block === 0) {
-		damage += bonusDamage;
+		damage += bonus;
 	}
 	if (user.element === element) {
 		addModifier(target, elementStagger);