+++ Source/equipment/potionkit-guarding.js
@@ -3,11 +3,11 @@ const Resource = require('../../Classes/Resource.js');
 const { removeModifier, getFullName, addBlock } = require('../combatantDAO.js');
 const { rollConsumable } = require('../labyrinths/_labyrinthDictionary');
 
-module.exports = new EquipmentTemplate("Guarding Potion Kit", "Gain @{block} block and add 1 random potion to loot", "Instead add @{critBonus} potions", "Water", effect, ["Urgent Potion Kit"])
+module.exports = new EquipmentTemplate("Guarding Potion Kit", "Gain @{block} block and add 1 random potion to loot", "Instead add @{critBonus} potions", "Water", effect, ["Organic Potion Kit", "Urgent Potion Kit"])
 	.setCategory("Trinket")
 	.setTargetingTags({ target: "none", team: "none" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }])
-	.setCost(200)
+	.setCost(350)
 	.setUses(10)
 	.setBlock(75);