+++ Source/equipment/shortsword-base.js
@@ -1,0 +1,29 @@
+const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
+const { dealDamage, addModifier, getFullName } = require("../combatantDAO.js");
+
+module.exports = new EquipmentTemplate("Shortsword", "Strike a foe for @{damage} @{element} damage, then apply @{mod1Stacks} @{mod1} to both the foe and yourself", "Damage x@{critBonus}", "Fire", effect, ["Accelerating Shortsword", "Toxic Shortsword"])
+	.setCategory("Weapon")
+	.setTargetingTags({ target: "single", team: "enemy" })
+	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Exposed", stacks: 1 }])
+	.setCost(200)
+	.setUses(10)
+	.setDamage(75);
+
+function effect([target], user, isCrit, adventure) {
+	if (target.hp < 1) {
+		return ` ${getFullName(target, adventure.room.enemyTitles)} was already dead!`;
+	}
+
+	let { element, modifiers: [elementStagger, exposed], damage, critBonus } = module.exports;
+	if (user.element === element) {
+		addModifier(target, elementStagger);
+	}
+	if (isCrit) {
+		damage *= critBonus;
+	}
+	addModifier(user, exposed);
+	return dealDamage(target, user, damage, false, element, adventure).then(resultText => {
+		addModifier(target, exposed);
+		return resultText;
+	});
+}