+++ Source/equipment/spear-reactive.js
@@ -1,7 +1,7 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier, calculateTotalSpeed, getFullName } = require("../combatantDAO.js");
 
-module.exports = new EquipmentTemplate("Reactive Spear", "Strike a foe for @{damage} (+@{bonus} if foe went first) @{element} damage", "Also inflict @{mod1Stacks} @{mod1}", "Fire", effect, ["Lethal Spear", "Sweeping Spear"])
+module.exports = new EquipmentTemplate("Reactive Spear", "Strike a foe for @{damage} (+@{bonus} if foe went first) @{element} damage", "Also inflict @{mod1Stacks} @{mod1}", "Wind", effect, ["Lethal Spear", "Sweeping Spear"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Stagger", stacks: 1 }])
@@ -15,9 +15,9 @@ function effect([target], user, isCrit, adventure) {
 		return ` ${getFullName(target, adventure.room.enemyTitles)} was already dead!`;
 	}
 
-	let { element, modifiers: [elementStagger, critStagger], damage, bonusDamage } = module.exports;
+	let { element, modifiers: [elementStagger, critStagger], damage, bonus } = module.exports;
 	if (calculateTotalSpeed(target) > calculateTotalSpeed(user)) {
-		damage += bonusDamage;
+		damage += bonus;
 	}
 	if (user.element === element) {
 		addModifier(target, elementStagger);